changed the way load logo, added verification when del items category,modified some views minor changes

+++ application/controllers/Item_category.php
@@ -19,10 +19,10 @@ class item_category extends CI_Controller {
 		$seller_id = $this->waimai_seller->check_login();
 		$this->isloadtemplate = 0;		
 		$data = array(
-		    'static_base_url' => $this->config->item('static_base_url'),
-		    'seller_name' => $this->session->seller_name,
-		    'logo_url' => ($this->cache->get('logo_url'.$seller_id)!='uploads/')?($this->cache->get('logo_url'.$seller_id)):"",
-		);
+			'static_base_url' => $this->config->item('static_base_url'),
+			'seller_name' => $this->session->seller_name,
+			'logo_url' => ($this->session->seller_logo_url != 'uploads/')? $this->session->seller_logo_url :"",
+			);
 		$data['validation_errors'] = '';
 		$data['result_success'] = '';
 		
@@ -31,51 +31,65 @@ class item_category extends CI_Controller {
 		$this->load->database();
 		if (empty($post_data))
 		{
-				$datalist = array();
-				$query = $this->db->query("SELECT * FROM ".$this->db->dbprefix('item_category')." WHERE seller_id=$seller_id and (parent_id=0 or isnull(parent_id)) ORDER BY id ASC");
-				$query_array = $query->result_array();
-				array_multisort(array_column($query_array, 'weight'), SORT_ASC, $query_array);
+			$datalist = array();
+			$query = $this->db->query("SELECT * FROM ".$this->db->dbprefix('item_category')." WHERE seller_id=$seller_id and (parent_id=0 or isnull(parent_id)) ORDER BY id ASC");
+			$query_array = $query->result_array();
+			array_multisort(array_column($query_array, 'weight'), SORT_DESC, $query_array);
 				//var_dump($query_array);
 
-				foreach ($query_array as $row)
-				{
-						$datalist[] = $row;
-				}
-				$query_total = $this->db->query("SELECT count(*) as total FROM ".$this->db->dbprefix('item_category')." WHERE seller_id=$seller_id and (parent_id=0 or isnull(parent_id)) ORDER BY id ASC");
-				 
-				$data['data_list'] = $datalist;	
-				$data['total']	= $query_total->row_array()['total'];
+			foreach ($query_array as $row)
+			{
+				$datalist[] = $row;
+			}
+			$query_total = $this->db->query("SELECT count(*) as total FROM ".$this->db->dbprefix('item_category')." WHERE seller_id=$seller_id and (parent_id=0 or isnull(parent_id)) ORDER BY id ASC");
+
+			$data['data_list'] = $datalist;	
+			$data['total']	= $query_total->row_array()['total'];
 				//var_dump($data['total']);	
-				$this->isloadtemplate = 1;
+			$this->isloadtemplate = 1;
 		}else
 		{
-				if ($post_data['ac'] == 'delete')
-				{
-						$id = $post_data['id'];						
-						$w = array(
-								'seller_id' => $seller_id,
-								'id' => $id,
+			if ($post_data['ac'] == 'delete')
+			{
+
+				$id = $post_data['id'];	
+				$query1 = $this->db->query("SELECT count(*) as total FROM ".$this->db->dbprefix('item')." WHERE seller_id=$seller_id and category_id = $id LIMIT 1");
+				
+				$is_product_category =  (int) $query1->row_array()['total'];
+				if($is_product_category == 0)
+				{	
+
+					$w = array(
+						'seller_id' => $seller_id,
+						'id' => $id,
 						);
-						$this->db->delete($this->db->dbprefix('item_category'),$w);
-						$result = $this->db->affected_rows();
-						if ($result)
-						{
-								$msg = 'ok';
-								$content = '';
-						}else
-						{
-								$msg = 'error';
-								$content = '';
-						}
-						$arr = array("msg" => $msg,"content" => $content);
-						$return_str = json_encode($arr);
-						exit($return_str);
+					$this->db->delete($this->db->dbprefix('item_category'),$w);
+					$result = $this->db->affected_rows();
+					if ($result)
+					{
+						$msg = 'ok';
+						$content = 'success';
+					}else
+					{
+						$msg = 'error';
+						$content = 'cant delete the category';
+					}
+					
 				}
+				else {
+					$msg = 'error';
+					$content = 'fail, certain products are under this category';
+				}
+
+				$arr = array("msg" => $msg,"content" => $content);
+				$return_str = json_encode($arr);
+				exit($return_str);
+			}
 		}
 		$this->db->close();
 		if($this->isloadtemplate)
 		{
-				$this->parser->parse('item_category_template', $data);
+			$this->parser->parse('item_category_template', $data);
 		}
 	}
